Plugin: update tests
